+++ build_onload_image.rb
@@ -5,6 +5,7 @@
 require 'getoptlong'
 
 ONLOAD_VERSIONS = {
+    '7.1.2.141'   => { :version => '7.1.2.141',   :md5sum => 'bfda4a68267e2aa3d5bed02af229b4fc', :driverid => '1d52732765feca797791b9668b14fb4e', :package_url => "https://support-nic.xilinx.com/wp/onload?sd=SF-109585-LS-36&pe=SF-122921-DH-5" },
     '7.1.1.75'    => { :version => '7.1.1.75',    :md5sum => '39b2d8d40982f6f3afd3cdb084969e90', :driverid => '65869c81c4a7f92b75316cf88446a9f1', :package_url => "https://support-nic.xilinx.com/wp/onload?sd=SF-109585-LS-35&pe=SF-122921-DH-4" },
     '7.1.0.265'   => { :version => '7.1.0.265',   :md5sum => '4db72fe198ec88d71fb1d39ef60c5ba7', :driverid => 'd9857bc9bddb5c6abdeb3f22d69b21d1', :package_url => "https://support.solarflare.com/wp/onload?sd=SF-109585-LS-33&pe=SF-122921-DH-2" },
     '7.0.0.176'   => { :version => '7.0.0.176',   :md5sum => '851ccf4fc76c96bcbeb01e1c57b20cce', :driverid => '6ac17472788a64c61013f3d7ed9ae4c9', :package_url => "https://support.solarflare.com/wp/onload?sd=SF-109585-LS-32&pe=SF-122921-DH-2" },
@@ -25,6 +26,7 @@ ONLOAD_VERSIONS = {
 
 IMAGE_FLAVORS = {
     'bionic'   => { :flavor => 'bionic' },
+    'bullseye' => { :flavor => 'bullseye' },
     'buster'   => { :flavor => 'buster' },
     'centos7'  => { :flavor => 'centos7' },
     'centos8'  => { :flavor => 'centos8' },