+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  * Update to Onload 7.1.1.75
  * Only build 64-bit userspace components (no 32-bit applications)
+ * `focal` flavor builds with GCC 10
  * Archive `precise` flavor
 
 ## 2020-10-31 (tagged 7.1.0.265)